+++ src/models/dynamic_generator.jl
@@ -6,7 +6,7 @@ mutable struct DynamicGenerator{
     P<:PSS,
 } <: DynamicInjection
     number::Int64
-    name::Symbol
+    name::String
     bus::Bus
     ω_ref::Float64
     V_ref::Float64
@@ -18,9 +18,10 @@ mutable struct DynamicGenerator{
     pss::P
     n_states::Int64
     states::Vector{Symbol}
+    ext::Dict{String, Any}
     function DynamicGenerator(
         number::Int64,
-        name::Symbol,
+        name::String,
         bus::Bus,
         ω_ref::Float64,
         V_ref::Float64,
@@ -30,7 +31,7 @@ mutable struct DynamicGenerator{
         avr::A,
         tg::TG,
         pss::P,
-    ) where {M<:Machine,S<:Shaft,A<:AVR,TG<:TurbineGov,P<:PSS}
+    ) where {M<:Machine, S<:Shaft, A<:AVR, TG<:TurbineGov, P<:PSS}
 
         n_states = (machine.n_states + shaft.n_states + avr.n_states + tg.n_states +
                     pss.n_states)
@@ -51,9 +52,12 @@ mutable struct DynamicGenerator{
             pss,
             n_states,
             states,
+            Dict{String, Any}()
         )
     end
 end
 
 get_Sbase(device::DynamicGenerator) = device.machine.s_rated
 get_Vref(device::DynamicGenerator) = device.V_ref
+get_bus(device::DynamicGenerator) = device.bus
+get_ext(device::DynamicGenerator) = device.ext